requirements.txt
pillow==11.1.0
tqdm==4.67.1
pyyaml==6.0.2